Added a method to check for tags assoicated with host

+++ isidore-stg/plugins/module/hosts.py
@@ -82,20 +82,26 @@ def run_module():
     elif module.params['state'] == 'absent':
         host = isidore.getHost(module.params['name'])
         if host:
-            if not module.check_mode:
-                # First, delete the associated tags from the HostHasTag table
-                # This is a test!
-                cursor = isidore._conn.cursor()
-                stmt_remove_tags = "DELETE FROM HostHasTag WHERE HostID = %s"
-                cursor.execute(stmt_remove_tags, [host._hostId])
-                cursor.close()
-
-                host.delete()
-            result['changed'] = True
-            result['message'] = 'Host was deleted.'
+            associated_tags = host.getTags()  # Check for associated tags
+            if associated_tags:
+                result['changed'] = False
+                result['message'] = 'Cannot delete host. Please remove the association of tags to the host before removing it.'
+            else:
+                if not module.check_mode:
+                    host.delete()
+                result['changed'] = True
+                result['message'] = 'Host was deleted.'
         else:
             result['changed'] = False
             result['message'] = 'Host does not exist.'
+        # if host:
+        #     if not module.check_mode:
+        #         host.delete()
+        #     result['changed'] = True
+        #     result['message'] = 'Host was deleted.'
+        # else:
+        #     result['changed'] = False
+        #     result['message'] = 'Host does not exist.'
 
 
     module.exit_json(**result)
